requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
distro===None
fastapi==0.120.0
firebase_admin==7.1.0
google-api-core==2.27.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-speech==2.34.0
google-cloud-storage==3.4.1
google-cloud-texttospeech==2.33.0
google-crc32c==1.7.1
google-resumable-media===None
googleapis-common-protos==1.71.0
grpcio===None
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack===None
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe===None
idna==3.11
jiter==0.11.1
msgpack==1.1.2
openai==2.6.1
proto-plus===None
protobuf===None
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser===None
pydantic==2.12.3
pydantic_core==2.41.4
pyjwt===None
python-dateutil===None
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six===None
sniffio===None
starlette==0.48.0
tqdm===None
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets===None
